fix: publish canary via pkg-pr-new's own --pnpm packing instead of prebuilt tarballs
